% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requal.R
\name{\%==\%}
\alias{\%==\%}
\title{Test rough equality}
\usage{
values \%==\% testval
}
\arguments{
\item{values}{close are any number of floating values to testval}

\item{testval}{value to be tested}

\item{epsilon}{float value indicating how large difference between values is allowed to be}
}
\value{

}
\description{
Test rough equality
}
\examples{
c(0.01,0.02,0.011)\%==\%0.01 equals TRUE FALSE  TRUE
}
