% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps2sp.R
\name{maps2sp}
\alias{maps2sp}
\title{Function to extract spatial polygon for subset of map}
\usage{
maps2sp(
  xlim,
  ylim,
  l.out = 100,
  clip = TRUE,
  crs = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
)
}
\arguments{
\item{crs}{}
}
\value{

}
\description{
This Function was adapted from https://edzer.github.io/sp/
}
