% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metplot.R
\name{metplot}
\alias{metplot}
\title{metplot}
\usage{
metplot(field, lons, lats, pbmin, pbmax, pbby, zeroinmid, guide.title, ...)
}
\arguments{
\item{field}{field to be plotted}

\item{lons}{vector of longitudes}

\item{lats}{vector of latitudes}

\item{pbmin}{minimum plotting level}

\item{pbmax}{maximum plotting level}

\item{pbby}{delta plotting level}

\item{zeroinmid}{should center be white?}

\item{guide.title}{guide title}
}
\value{
plot
}
\description{
metplot
}
\examples{
metplot(APVTOT,lons=lons,lats=lats,pbmin=-40,pbmax=40,pbby=5,zeroinmid=F,guide.title="APV (PVU)")
}
