% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkrast.R
\name{mkrast}
\alias{mkrast}
\title{Make raster from field}
\usage{
mkrast(
  field,
  lonmin = -180,
  lonmax = 180,
  latmin = -90,
  latmax = 90,
  dellon = 1,
  dellat = 1,
  crsstring = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
  ...
)
}
\arguments{
\item{field}{this is the field...}

\item{dellat}{}
}
\value{

}
\description{
Make raster from field
}
\examples{
mkrast(FIELD)
}
