% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfile.R
\name{rfile}
\alias{rfile}
\title{Read NetCDF File}
\usage{
rfile(file, field, ...)
}
\arguments{
\item{file}{character string specifying the NetCDF file location}

\item{field}{character string specifying the field to be read}

\item{...}{arbitrary number of additional character strings specifying additional fields to be read}
}
\value{

}
\description{
Read NetCDF File
}
\examples{
readfile("/net/server/folder/somefile.nc","SLP","OMEGA")
}
