% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spldflen.R
\name{spldflen}
\alias{spldflen}
\title{Length of SpatialLinesDataFrames Elements}
\usage{
spldflen(spldf)
}
\arguments{
\item{spldf}{A SpatialLinesDataFrame with an arbitrary number of Lines}
}
\value{

}
\description{
Length of SpatialLinesDataFrames Elements
}
\examples{
spldf(x)
}
