% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metplot.R
\name{metplot}
\alias{metplot}
\title{metplot}
\usage{
metplot(field, lonmin, lonmax, latmin, latmax, level)
}
\arguments{
\item{field}{which field should be plottet}

\item{lonmin}{minimum longitude}

\item{lonmax}{max lon}

\item{latmin}{min lat}

\item{latmax}{max lat}

\item{level}{model level}
}
\value{
plot
}
\description{
metplot
}
\examples{
metplot("/home/atroman/phd/file.nc",1)
}
