% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkrast.R
\name{mkrast}
\alias{mkrast}
\title{Make raster from field}
\usage{
mkrast(
  field,
  lonmin = -180,
  lonmax = 180,
  latmin = -90,
  latmax = 90,
  dellon = 1,
  dellat = 1,
  crsstring = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
  lons = NA,
  lats = NA
)
}
\arguments{
\item{field}{A matrix or array of the field to be converted into a raster object}

\item{lonmin}{A float specifying min longitude}

\item{lonmax}{A float specifying max longitude}

\item{latmin}{A float specifying min latitude}

\item{latmax}{A float specifying max latitude}

\item{dellon}{A float specifying the delta longitude}

\item{dellat}{A float specifying the delta latitude}

\item{crsstring}{A string specifying the projection to be used}

\item{lons}{A vector of longitudes}

\item{lats}{A vector of latitudes}
}
\value{

}
\description{
Make raster from field
}
\examples{
PVr <- mkrast(PV,lons=lons,lats=lats)
}
