% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfile.R
\name{rfile}
\alias{rfile}
\title{Read NetCDF File}
\usage{
rfile(file, field, ..., navals = -999)
}
\arguments{
\item{file}{character string specifying the NetCDF file location}

\item{field}{character string specifying the field to be read}

\item{...}{arbitrary number of character strings specifying additional fields to be read}

\item{navals}{float specifying NA values}
}
\value{

}
\description{
Read NetCDF File
}
\examples{
library(zeallot) # Need this to unpack data from rfile
c(APVTURBT)\%<-\%rfile("/net/thermo/atmosdyn2/atroman/stingjet/20140211_12/tra/APV12_20140212_07.nc","APVTURBT")
}
